dimension fixes

+++ quantumfrqi/quantumfrqi.py
@@ -35,9 +35,9 @@ def processimage(image_name, num_shots=1024*1024, side_dimension=16):
     print(f"Processing {image_name} with {num_shots} shots...")
     if(min(og_width, og_height) > side_dimension):
         if og_width >= og_height: 
-            SIZE = (side_dimension, side_dimension*og_height/og_width)
+            SIZE = (side_dimension, int(side_dimension*og_height/og_width))
         if og_width > og_height: 
-            SIZE = (side_dimension*og_width/og_height, side_dimension)
+            SIZE = (int(side_dimension*og_width/og_height), side_dimension)
         prepare_dimensions(f"./data/images/{IMAGE_NAME}.png", f"./data/images/{IMAGE_NAME}{side_dimension}.png", SIZE)
         IMAGE_NAME = IMAGE_NAME+str(side_dimension)
     image_path = f"./data/images/{IMAGE_NAME}.png"
